Move split algorithms out of constrained basis
