+++ openghg_inversions/basis/algorithms/_consolidation.py
@@ -21,8 +21,7 @@ from ._contrast import (
 )
 
 if TYPE_CHECKING:
-    from ._constrained import LatLonGridGeometry
-    from ._partition import GridNode
+    from ._partition import GridNode, LatLonGridGeometry
 
 InactiveComponentPolicy = Literal["keep", "merge_nearest"]
